pyproject.toml
[tool.poetry]
name = "fastapi-ci"
version = "0.1.0"
description = "Testing FastAPI application with Gitlab CI"
authors = ["Oleksii-Op <oleksii.oprish@gmail.com>"]
license = "GPLv2"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic-settings = "^2.7.1"
sqlmodel = "^0.0.22"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
email-validator = "^2.2.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.4"
mypy = "^1.14.1"
faker = "^35.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
